NLP/fnlp-assignment2/requirements.txt
torch
pytorch-crf
pandas